split OtherFeats into ClassFeats
