+++ SolastaUnfinishedBusiness/Feats/OtherFeats.cs
@@ -31,8 +31,6 @@ internal static class OtherFeats
     internal static void CreateFeats([NotNull] List<FeatDefinition> feats)
     {
         var featAstralArms = BuildAstralArms();
-        var featCallForCharge = BuildCallForCharge();
-        var featCunningEscape = BuildCunningEscape();
         var featEldritchAdept = BuildEldritchAdept();
         var featHealer = BuildHealer();
         var featInspiringLeader = BuildInspiringLeader();
@@ -41,7 +39,6 @@ internal static class OtherFeats
         var featMonkInitiate = BuildMonkInitiate();
         var featPickPocket = BuildPickPocket();
         var featPoisonousSkin = BuildPoisonousSkin();
-        var featPotentSpellcaster = BuildPotentSpellcaster();
         var featTough = BuildTough();
         var featWarCaster = BuildWarcaster();
 
@@ -52,8 +49,6 @@ internal static class OtherFeats
 
         feats.AddRange(
             featAstralArms,
-            featCallForCharge,
-            featCunningEscape,
             featEldritchAdept,
             featHealer,
             featInspiringLeader,
@@ -62,7 +57,6 @@ internal static class OtherFeats
             featMonkInitiate,
             featPickPocket,
             featPoisonousSkin,
-            featPotentSpellcaster,
             featTough,
             featWarCaster);
 
@@ -72,14 +66,19 @@ internal static class OtherFeats
             featPoisonousSkin);
 
         GroupFeats.FeatGroupSupportCombat.AddFeats(
-            featCallForCharge,
             featHealer,
             featInspiringLeader);
 
         GroupFeats.FeatGroupAgilityCombat.AddFeats(
-            featCunningEscape,
             featMobile);
 
+        GroupFeats.FeatGroupSpellCombat.AddFeats(
+            FeatDefinitions.FlawlessConcentration,
+            FeatDefinitions.PowerfulCantrip,
+            elementalAdeptGroup,
+            featWarCaster,
+            spellSniperGroup);
+        
         GroupFeats.MakeGroup("FeatGroupBodyResilience", null,
             FeatDefinitions.BadlandsMarauder,
             FeatDefinitions.BlessingOfTheElements,
@@ -94,71 +93,7 @@ internal static class OtherFeats
             FeatDefinitions.Manipulator,
             featHealer,
             featPickPocket);
-
-        GroupFeats.MakeGroup("FeatGroupSpellCombat", null,
-            FeatDefinitions.FlawlessConcentration,
-            FeatDefinitions.PowerfulCantrip,
-            elementalAdeptGroup,
-            featPotentSpellcaster,
-            featWarCaster,
-            spellSniperGroup);
-
-        GroupFeats.MakeGroup("FeatGroupClassBound", null,
-            featCallForCharge,
-            featCunningEscape,
-            featPotentSpellcaster);
     }
-
-    #region Call for Charge
-
-    private static FeatDefinition BuildCallForCharge()
-    {
-        const string NAME = "FeatCallForCharge";
-
-        return FeatDefinitionWithPrerequisitesBuilder
-            .Create("FeatCallForCharge")
-            .SetGuiPresentation(Category.Feat)
-            .SetFeatures(FeatureDefinitionPowerBuilder
-                .Create($"Power{NAME}")
-                .SetGuiPresentation(Category.Feature, PowerOathOfTirmarGoldenSpeech)
-                .SetUsesAbilityBonus(ActivationTime.BonusAction, RechargeRate.LongRest, AttributeDefinitions.Charisma)
-                .SetEffectDescription(
-                    EffectDescriptionBuilder
-                        .Create()
-                        .SetTargetingData(Side.Ally, RangeType.Self, 0, TargetType.Sphere, 6)
-                        .SetDurationData(DurationType.Round, 1)
-                        .SetEffectForms(
-                            EffectFormBuilder
-                                .Create()
-                                .SetConditionForm(
-                                    ConditionDefinitionBuilder
-                                        .Create($"Condition{NAME}")
-                                        .SetGuiPresentation(Category.Condition, ConditionDefinitions.ConditionBlessed)
-                                        .SetSpecialInterruptions(ConditionInterruption.Attacked)
-                                        .SetPossessive()
-                                        .SetFeatures(
-                                            FeatureDefinitionMovementAffinityBuilder
-                                                .Create($"MovementAffinity{NAME}")
-                                                .SetGuiPresentation($"Condition{NAME}", Category.Condition)
-                                                .SetBaseSpeedAdditiveModifier(3)
-                                                .AddToDB(),
-                                            FeatureDefinitionCombatAffinityBuilder
-                                                .Create($"CombatAffinity{NAME}")
-                                                .SetGuiPresentation($"Condition{NAME}", Category.Condition)
-                                                .SetMyAttackAdvantage(AdvantageType.Advantage)
-                                                .AddToDB())
-                                        .AddToDB(),
-                                    ConditionForm.ConditionOperation.Add)
-                                .Build())
-                        .SetParticleEffectParameters(SpellDefinitions.MagicWeapon)
-                        .Build())
-                .AddToDB())
-            .SetAbilityScorePrerequisite(AttributeDefinitions.Charisma, 13)
-            .SetValidators(ValidatorsFeat.IsPaladinLevel1)
-            .AddToDB();
-    }
-
-    #endregion
 
     #region Eldritch Adept
 
@@ -490,40 +425,6 @@ internal static class OtherFeats
 
             attackMode.reach = true;
             attackMode.reachRange = 2;
-        }
-    }
-
-    #endregion
-
-
-    #region Cunning Escape
-
-    private static FeatDefinition BuildCunningEscape()
-    {
-        return FeatDefinitionWithPrerequisitesBuilder
-            .Create("FeatCunningEscape")
-            .SetGuiPresentation(Category.Feat)
-            .SetCustomSubFeatures(new OnAfterActionFeatureFeatCunningEscape())
-            .SetValidators(ValidatorsFeat.IsRogueLevel3)
-            .AddToDB();
-    }
-
-    private class OnAfterActionFeatureFeatCunningEscape : IOnAfterActionFeature
-    {
-        public void OnAfterAction(CharacterAction action)
-        {
-            if (action.ActionDefinition != DatabaseHelper.ActionDefinitions.DashBonus)
-            {
-                return;
-            }
-
-            var actingCharacter = action.ActingCharacter.RulesetCharacter;
-
-            var condition = RulesetCondition.CreateActiveCondition(
-                actingCharacter.Guid, ConditionDefinitions.ConditionDisengaging, DurationType.Round, 0,
-                TurnOccurenceType.EndOfTurn, actingCharacter.Guid, actingCharacter.CurrentFaction.Name);
-
-            actingCharacter.AddConditionOfCategory(AttributeDefinitions.TagCombat, condition);
         }
     }
 
@@ -871,50 +772,6 @@ internal static class OtherFeats
             var effectPower = GetUsablePower(attacker);
 
             effectPower.ApplyEffectOnCharacter(defender.RulesetCharacter, true, defender.LocationPosition);
-        }
-    }
-
-    #endregion
-
-    #region Potent Spellcaster
-
-    private static FeatDefinition BuildPotentSpellcaster()
-    {
-        return FeatDefinitionWithPrerequisitesBuilder
-            .Create("FeatPotentSpellcaster")
-            .SetGuiPresentation(Category.Feat)
-            .SetCustomSubFeatures(new ModifyMagicEffectFeatPotentSpellcaster())
-            .SetValidators(ValidatorsFeat.IsWizardLevel6)
-            .AddToDB();
-    }
-
-    private sealed class ModifyMagicEffectFeatPotentSpellcaster : IModifyMagicEffect
-    {
-        public EffectDescription ModifyEffect(
-            BaseDefinition definition,
-            EffectDescription effect,
-            RulesetCharacter character)
-        {
-            if (definition is not SpellDefinition spellDefinition ||
-                !SpellListDefinitions.SpellListWizard.SpellsByLevel.Any(x =>
-                    x.Level == 0 && x.Spells.Contains(spellDefinition)))
-            {
-                return effect;
-            }
-
-            var damage = effect.FindFirstDamageForm();
-
-            if (damage == null)
-            {
-                return effect;
-            }
-
-            damage.BonusDamage += AttributeDefinitions.ComputeAbilityScoreModifier(
-                character.GetAttribute(AttributeDefinitions.Intelligence).CurrentValue);
-            damage.DamageBonusTrends.Add(
-                new TrendInfo(1, FeatureSourceType.CharacterFeature, "Feat/&FeatPotentSpellcasterTitle", null));
-
-            return effect;
         }
     }
