+++ SolastaUnfinishedBusiness/Translations/en/Feats/ClassFeats-en.txt
@@ -1,0 +1,10 @@
+Condition/&ConditionFeatCallForChargeDescription=Your speed increase by 15 ft and you have advantage on your attack.
+Condition/&ConditionFeatCallForChargeTitle=Call for Charge
+Feat/&FeatCallForChargeDescription=You may use your bonus action to call for a charge. When you do, until the start of your next turn, you and your allies speed increase by 15 ft, and each one have advantage on their first attack roll. You can use this power a number of times per long rest equal to your Charisma modifier.
+Feat/&FeatCallForChargeTitle=Call for Charge
+Feat/&FeatCunningEscapeDescription=Whenever you use the Dash action as a bonus action, your movement doesn't provoke opportunity attacks for the rest of the turn.
+Feat/&FeatCunningEscapeTitle=Cunning Escape
+Feat/&FeatPotentSpellcasterDescription=You can add your Intelligence modifier to the damage you deal with any wizard cantrip.
+Feat/&FeatPotentSpellcasterTitle=Potent Spellcaster
+Feature/&PowerFeatCallForChargeDescription=Your and your allies speed increase by 15 ft and you have advantage on your attack.
+Feature/&PowerFeatCallForChargeTitle=Call for Charge
