+++ SolastaUnfinishedBusiness/Translations/en/Feats/OtherFeats-en.txt
@@ -1,13 +1,7 @@
-Condition/&ConditionFeatCallForChargeDescription=Your speed increase by 15 ft and you have advantage on your attack.
-Condition/&ConditionFeatCallForChargeTitle=Call for Charge
 Condition/&ConditionFeatMobileAfterDashDescription=Grants immunity to movement restrictions as well as being restrained.
 Condition/&ConditionFeatMobileAfterDashTitle=Freedom
 Feat/&FeatAstralArmsDescription=Increase your Wisdom by 1, to a maximum of 20.\nWhen you make an unarmed strike on your turn, your reach for it is 10 feet.
 Feat/&FeatAstralArmsTitle=Astral Reach
-Feat/&FeatCallForChargeDescription=You may use your bonus action to call for a charge. When you do, until the start of your next turn, you and your allies speed increase by 15 ft, and each one have advantage on their first attack roll. You can use this power a number of times per long rest equal to your Charisma modifier.
-Feat/&FeatCallForChargeTitle=Call for Charge
-Feat/&FeatCunningEscapeDescription=Whenever you use the Dash action as a bonus action, your movement doesn't provoke opportunity attacks for the rest of the turn.
-Feat/&FeatCunningEscapeTitle=Cunning Escape
 Feat/&FeatEldritchAdeptDescription=You learn one Eldritch Invocation option of your choice from the warlock class. If the invocation has a prerequisite, you can choose that invocation only if you're a warlock and only if you meet the prerequisite. Whenever you gain a level, you can replace the invocation with another one from the warlock class.
 Feat/&FeatEldritchAdeptTitle=Eldritch Adept
 Feat/&FeatElementalAdeptDescription={0} damage type spells you cast ignore target's resistance to damage. In addition, when you roll damage for these spells, you can reroll any 1s.
@@ -28,8 +22,6 @@ Feat/&FeatPickPocketDescription=You gain proficiency with Sleight of Hand, or ex
 Feat/&FeatPickPocketTitle=Pickpocket
 Feat/&FeatPoisonousSkinDescription=Any creature that grapples you or otherwise comes into direct contact with your skin must succeed on a saving throw or becomes poisoned for 1 minute.
 Feat/&FeatPoisonousSkinTitle=Poisonous Skin
-Feat/&FeatPotentSpellcasterDescription=You can add your Intelligence modifier to the damage you deal with any wizard cantrip.
-Feat/&FeatPotentSpellcasterTitle=Potent Spellcaster
 Feat/&FeatSpellSniperDescription=You learn two cantrips of your choice from the {0} spell list. When you cast a spell that requires you to make an attack roll, the spell's range is doubled. Your ranged spell attacks ignore half cover and three-quarters cover.
 Feat/&FeatSpellSniperTitle=Spell Sniper [{0}]
 Feat/&FeatToughDescription=Your hit point maximum increases by an amount equal to twice your level when you gain this feat. Whenever you gain a level thereafter, your hit point maximum increases by an additional 2 hit points.
@@ -39,8 +31,6 @@ Feat/&FeatWarCasterTitle=War Caster
 Feature/&CastSpellFeatMagicInitiateTitle=Initiate
 Feature/&CastSpellFeatSpellSniperTitle=Sniper
 Feature/&DieRollModifierFeatElementalAdeptReroll=Because of Elemental Adept {0} re-rolls the {1} die from a {2} to a {3}.
-Feature/&PowerFeatCallForChargeDescription=Your and your allies speed increase by 15 ft and you have advantage on your attack.
-Feature/&PowerFeatCallForChargeTitle=Call for Charge
 Feature/&PowerFeatHealerMedKitDescription=You can use an action to restore 1d6 + 4 + character level hit points a number of times per short rest equal to your Wisdom modifier.
 Feature/&PowerFeatHealerMedKitTitle=Use Medicine
 Feature/&PowerFeatHealerResuscitateDescription=Use an action to revive a creature that died within the last 12 seconds.
